backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
PyJWT
cryptography
httpx
python-multipart

# LangChain for multi-LLM support (Claude, OpenAI, Gemini)
langchain
langchain-core
langchain-anthropic
langchain-openai
langchain-google-genai
